部分代码重构

+++ src/main/java/com/oven/controller/EmployeeController.java
@@ -1,8 +1,8 @@
 package com.oven.controller;
 
 import com.alibaba.fastjson.JSONObject;
-import com.oven.contants.AppConst;
-import com.oven.contants.PermissionCode;
+import com.oven.constant.AppConst;
+import com.oven.constant.PermissionCode;
 import com.oven.enumerate.ResultEnum;
 import com.oven.exception.MyException;
 import com.oven.service.EmployeeService;
@@ -73,7 +73,7 @@ public class EmployeeController extends BaseController {
                 item.setCreateName(userService.getById(item.getCreateId()).getNickName());
                 item.setLastModifyName(userService.getById(item.getLastModifyId()).getNickName());
             }
-            Long totalNum = employeeService.getTotalNum(employee);
+            Integer totalNum = employeeService.getTotalNum(employee);
             result.put("code", 0);
             result.put("msg", "");
             result.put("count", totalNum);
